Continue CLI cleanup and share discovery helpers
