+++ src/cli.rs
@@ -1,5 +1,8 @@
 #[cfg(any(feature = "build", feature = "formatter", feature = "run", feature = "serve", feature = "bundle_web", feature = "test", feature = "repl"))]
 pub mod app;
+pub mod commands;
+pub(crate) mod paths;
+pub(crate) mod source_discovery;
 #[cfg(feature = "bundle_web")]
 pub mod bundle_web;
 #[cfg(any(feature = "serve", feature = "run"))]
