+++ src/cli/commands/mod.rs
@@ -1,0 +1,14 @@
+#[cfg(feature = "build")]
+pub(crate) mod build;
+#[cfg(feature = "formatter")]
+pub(crate) mod format;
+#[cfg(feature = "run")]
+pub(crate) mod run;
+#[cfg(feature = "serve")]
+pub(crate) mod serve;
+#[cfg(feature = "test")]
+pub(crate) mod test;
+#[cfg(feature = "bundle_web")]
+pub(crate) mod bundle_web;
+#[cfg(feature = "repl")]
+pub(crate) mod repl;
