+++ src/cli/commands/run.rs
@@ -1,0 +1,6 @@
+use clap::ArgMatches;
+use mech_core::*;
+
+pub(crate) fn run(_root_matches: &ArgMatches, _subcommand_matches: Option<&ArgMatches>) -> MResult<()> {
+  Ok(())
+}
